% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fundeb.R
\name{build_fundeb_url}
\alias{build_fundeb_url}
\title{Build FUNDEB distribution URL from lookup table}
\usage{
build_fundeb_url(year)
}
\arguments{
\item{year}{The year.}
}
\value{
A character string with the download URL.
}
\description{
Internal function to construct the download URL for FUNDEB distribution
data from STN. Uses a hardcoded lookup table of publication IDs.
}
\keyword{internal}
