% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ideb.R
\name{build_ideb_url}
\alias{build_ideb_url}
\title{Build IDEB download URL}
\usage{
build_ideb_url(level, stage, year)
}
\arguments{
\item{level}{Geographic level.}

\item{stage}{Education stage.}

\item{year}{The file year (most recent available).}
}
\value{
A character string with the URL.
}
\description{
Internal function to construct IDEB download URLs based on level and stage.
}
\keyword{internal}
