% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{build_inep_url}
\alias{build_inep_url}
\title{Build INEP microdata URL}
\usage{
build_inep_url(dataset, year, ...)
}
\arguments{
\item{dataset}{The dataset name (e.g., "censo_escolar", "enem").}

\item{year}{The year of the data.}

\item{...}{Additional parameters for URL construction.}
}
\value{
A character string with the URL.
}
\description{
Internal function to construct URLs for INEP microdata.
}
\keyword{internal}
