% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-saeb.R
\name{build_saeb_zip_filename}
\alias{build_saeb_zip_filename}
\title{Build SAEB ZIP filename}
\usage{
build_saeb_zip_filename(year, level = "fundamental_medio")
}
\arguments{
\item{year}{The year.}

\item{level}{The level (only relevant for 2021).}
}
\value{
The ZIP filename.
}
\description{
Internal function to build the correct ZIP filename for SAEB data.
Handles the special case of 2021 (split into two files).
}
\keyword{internal}
