% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cpc.R
\name{convert_faixa_columns}
\alias{convert_faixa_columns}
\title{Convert faixa columns to numeric}
\usage{
convert_faixa_columns(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The data frame with faixa columns as numeric.
}
\description{
Internal function to convert columns ending in \verb{_faixa} from character
to numeric. Values like \code{"SC"} (Sem Conceito) are converted to \code{NA}.
}
\keyword{internal}
