% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{detect_delim}
\alias{detect_delim}
\title{Detect file delimiter}
\usage{
detect_delim(file)
}
\arguments{
\item{file}{Path to the data file.}
}
\value{
The detected delimiter character.
}
\description{
Internal function to detect the delimiter used in a CSV file by reading
the first line and counting occurrences of common delimiters.
}
\keyword{internal}
