% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{detect_encoding}
\alias{detect_encoding}
\title{Detect file encoding}
\usage{
detect_encoding(file)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
A character string with the encoding name.
}
\description{
Internal function to detect the encoding of a text file.
INEP files typically use Latin-1 or UTF-8.
}
\keyword{internal}
