% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-escolar.R
\name{find_censo_file}
\alias{find_censo_file}
\title{Find the Censo Escolar data file}
\usage{
find_censo_file(exdir, year)
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the main data file within the
extracted census directory.
}
\keyword{internal}
