% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-superior.R
\name{find_censo_superior_file}
\alias{find_censo_superior_file}
\title{Find the Higher Education Census data file}
\usage{
find_censo_superior_file(exdir, year, type = "ies")
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}

\item{type}{The data type ("ies", "cursos", "alunos", "docentes").}
}
\value{
The path to the data file.
}
\description{
Internal function to locate a Higher Education Census data file within
the extracted directory based on the requested type.
}
\keyword{internal}
