% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-enade.R
\name{find_enade_file}
\alias{find_enade_file}
\title{Find the ENADE data file}
\usage{
find_enade_file(exdir, year)
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the ENADE data file within the extracted
directory.
}
\keyword{internal}
