% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-enem-escola.R
\name{find_enem_escola_file}
\alias{find_enem_escola_file}
\title{Find the ENEM por Escola data file}
\usage{
find_enem_escola_file(exdir)
}
\arguments{
\item{exdir}{The extraction directory.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the ENEM por Escola data file within
the extracted directory.
}
\keyword{internal}
