% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-idd.R
\name{find_idd_file}
\alias{find_idd_file}
\title{Find the IDD data file}
\usage{
find_idd_file(exdir, year)
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the IDD data file within the extracted
directory.
}
\keyword{internal}
