% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ideb.R
\name{fix_brasil_cols}
\alias{fix_brasil_cols}
\title{Fix brasil unnamed columns}
\usage{
fix_brasil_cols(df)
}
\arguments{
\item{df}{Data frame with brasil data.}
}
\value{
Data frame with renamed columns.
}
\description{
Internal function to rename columns in brasil data.
}
\keyword{internal}
