% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the current cache directory}
\usage{
get_cache_dir()
}
\value{
A character string with the path to the cache directory.
}
\description{
Returns the current cache directory used by educabR.
}
\examples{
get_cache_dir()
}
\seealso{
Other cache functions:
\code{\link[=clear_cache]{clear_cache()}},
\code{\link[=list_cache]{list_cache()}},
\code{\link[=set_cache_dir]{set_cache_dir()}}
}
\concept{cache functions}
