% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_afd.R
\name{le_afd}
\alias{le_afd}
\title{Processa arquivo XLSX de resultados da AFD com colunas hierárquicas}
\usage{
le_afd(
  ano = 2024,
  regiao = "municipios",
  localizacoes = "total",
  dependencias = "total",
  niveis = "ensino_medio",
  subniveis = "total"
)
}
\arguments{
\item{ano}{number , year}

\item{regiao}{character , one of escolas,municipios, ufs}

\item{localizacao}{character, one of "total","urbana","rural"}

\item{dependencia}{character, one of "total","publica","privada","estadual","municipal"}

\item{nivel}{character, one of infantil,ensino_fundamental,ensino_medio,eja_fundamental,eja_medio}

\item{subnivel}{character, one of total,anos_iniciais,anos_finais}
}
\value{
Tibble com dados formatados
}
\description{
Processa arquivo XLSX de resultados da AFD com colunas hierárquicas
}
