% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_censoescolar.R
\name{le_censoescolar}
\alias{le_censoescolar}
\title{Processa arquivo CSV e dicion XLSX de resultados do censoescolar com colunas hierárquicas - NÃO TERMINADO!}
\usage{
le_censoescolar(ano = 2023)
}
\arguments{
\item{ano}{character , 1997-2024}
}
\value{
Tibble com dados formatados
}
\description{
Processa arquivo CSV e dicion XLSX de resultados do censoescolar com colunas hierárquicas - NÃO TERMINADO!
}
