% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_idadeserie.R
\name{le_idadeserie}
\alias{le_idadeserie}
\title{Processa arquivo CSV e dicion XLSX de resultados do idadeserie com colunas hierárquicas - NÃO TERMINADO!}
\usage{
le_idadeserie(ano = 2023)
}
\arguments{
\item{ano}{character , 1997-2024}
}
\value{
Tibble com dados formatados
}
\description{
Processa arquivo CSV e dicion XLSX de resultados do idadeserie com colunas hierárquicas - NÃO TERMINADO!
}
