% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_ideb.R
\name{le_ideb}
\alias{le_ideb}
\title{Processa arquivo XLSX de resultados do IDEB com colunas hierárquicas}
\usage{
le_ideb(caminho_arquivo)
}
\arguments{
\item{caminho_arquivo}{Caminho para o arquivo .xlsx}
}
\value{
Tibble com dados formatados
}
\description{
Processa arquivo XLSX de resultados do IDEB com colunas hierárquicas
}
