% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{list_cache}
\alias{list_cache}
\title{List cached files}
\usage{
list_cache(dataset = NULL)
}
\arguments{
\item{dataset}{Optional. Filter by dataset name.}
}
\value{
A tibble with information about cached files.
}
\description{
Lists all files currently in the educabR cache.
}
\examples{
\dontrun{
list_cache()
}
}
\seealso{
Other cache functions:
\code{\link[=clear_cache]{clear_cache()}},
\code{\link[=get_cache_dir]{get_cache_dir()}},
\code{\link[=set_cache_dir]{set_cache_dir()}}
}
\concept{cache functions}
