% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-escolar.R
\name{parse_sas_dates}
\alias{parse_sas_dates}
\title{Parse SAS datetime columns to Date}
\usage{
parse_sas_dates(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The data frame with date columns converted.
}
\description{
Internal function to convert columns with SAS datetime format
(e.g. "12FEB2024:00:00:00") to Date objects.
}
\keyword{internal}
