% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cpc.R
\name{read_excel_safe}
\alias{read_excel_safe}
\title{Read Excel file safely}
\usage{
read_excel_safe(file, n_max = Inf)
}
\arguments{
\item{file}{Path to the Excel file.}

\item{n_max}{Maximum number of rows to read.}
}
\value{
A tibble with the data.
}
\description{
Internal function to read Excel files (xls/xlsx) with error handling.
Tries to read the first sheet by default.
}
\keyword{internal}
