% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{read_inep_file}
\alias{read_inep_file}
\title{Read INEP data file}
\usage{
read_inep_file(file, delim = ";", encoding = NULL, n_max = Inf, quiet = FALSE)
}
\arguments{
\item{file}{Path to the data file.}

\item{delim}{The delimiter character.}

\item{encoding}{The file encoding.}

\item{n_max}{Maximum number of rows to read.}

\item{quiet}{Logical. If \code{TRUE}, suppresses the large-file advisory.}
}
\value{
A tibble with the data.
}
\description{
Internal function to read INEP data files with appropriate settings.
}
\keyword{internal}
