% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-escolar.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize column names}
\usage{
standardize_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The data frame with standardized names.
}
\description{
Internal function to standardize column names to lowercase
with underscores.
}
\keyword{internal}
