% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-escolar.R
\name{uf_to_code}
\alias{uf_to_code}
\title{Convert UF abbreviation to code}
\usage{
uf_to_code(uf)
}
\arguments{
\item{uf}{UF abbreviation or code.}
}
\value{
The numeric UF code.
}
\description{
Internal function to convert state abbreviations to IBGE codes.
}
\keyword{internal}
