% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-download.R
\name{available_years}
\alias{available_years}
\title{Check available years for a dataset}
\usage{
available_years(dataset)
}
\arguments{
\item{dataset}{The dataset name.}
}
\value{
An integer vector of available years.
}
\description{
Returns the years available for a given INEP dataset.
}
\examples{
available_years("censo_escolar")
available_years("enem")
}
