% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the educabR cache}
\usage{
clear_cache(dataset = NULL)
}
\arguments{
\item{dataset}{Optional. A character string specifying which dataset
cache to clear. If \code{NULL}, clears all caches.}
}
\value{
Invisibly returns \code{TRUE} if successful.
}
\description{
Removes all cached files from the educabR cache directory.
}
\examples{
\donttest{
# clear all cached data
clear_cache()

# clear only ENEM cache
clear_cache("enem")
}
}
