% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-encceja.R
\name{find_encceja_file}
\alias{find_encceja_file}
\title{Find the ENCCEJA data file}
\usage{
find_encceja_file(exdir, year)
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the ENCCEJA data file within the extracted
directory.
}
\keyword{internal}
