% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-enem.R
\name{find_enem_file}
\alias{find_enem_file}
\title{Find the ENEM data file}
\usage{
find_enem_file(exdir, year)
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}
}
\value{
The path to the data file.
}
\description{
Internal function to locate the main ENEM data file within the
extracted directory.
}
\keyword{internal}
