% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-saeb.R
\name{find_saeb_file}
\alias{find_saeb_file}
\title{Find the SAEB data file}
\usage{
find_saeb_file(exdir, year, type = "aluno")
}
\arguments{
\item{exdir}{The extraction directory.}

\item{year}{The year.}

\item{type}{The data type ("aluno", "escola", "diretor", "professor").}
}
\value{
The path to the data file.
}
\description{
Internal function to locate a SAEB data file within the extracted
directory based on the requested type.
}
\keyword{internal}
