% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the current cache directory}
\usage{
get_cache_dir()
}
\value{
A character string with the path to the cache directory.
}
\description{
Returns the current cache directory used by educabR.
}
\examples{
get_cache_dir()
}
