% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-enem.R
\name{get_enem_itens}
\alias{get_enem_itens}
\title{Get ENEM item response data}
\usage{
get_enem_itens(year, n_max = Inf, quiet = FALSE)
}
\arguments{
\item{year}{The year of the exam (2009-2023).}

\item{n_max}{Maximum number of rows to read.}

\item{quiet}{Logical. If \code{TRUE}, suppresses progress messages.}
}
\value{
A tibble with item response data.
}
\description{
Downloads and processes ENEM item response (gabarito) data,
which contains detailed information about each question.
}
\examples{
\donttest{
# get item data for 2023
itens <- get_enem_itens(2023)
}
}
