% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_ideb.R
\name{le_ideb}
\alias{le_ideb}
\title{Processa arquivo XLSX de resultados do IDEB com colunas hierárquicas}
\usage{
le_ideb(regiao = "municipios", nivel = "iniciais")
}
\arguments{
\item{regiao}{character , one of municipios, ufs, brasil}

\item{nivel}{character, one of iniciais,finais,medio}
}
\value{
Tibble com dados formatados
}
\description{
Processa arquivo XLSX de resultados do IDEB com colunas hierárquicas
}
