% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{list_cache}
\alias{list_cache}
\title{List cached files}
\usage{
list_cache(dataset = NULL)
}
\arguments{
\item{dataset}{Optional. Filter by dataset name.}
}
\value{
A tibble with information about cached files.
}
\description{
Lists all files currently in the educabR cache.
}
\examples{
\donttest{
list_cache()
}
}
