% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-escolar.R
\name{list_censo_files}
\alias{list_censo_files}
\title{List available Censo Escolar files}
\usage{
list_censo_files(year)
}
\arguments{
\item{year}{The year of the census.}
}
\value{
A character vector of file names found.
}
\description{
Lists the data files available in a downloaded School Census.
}
\examples{
\donttest{
list_censo_files(2023)
}
}
