% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-censo-superior.R
\name{list_censo_superior_files}
\alias{list_censo_superior_files}
\title{List available Higher Education Census files}
\usage{
list_censo_superior_files(year)
}
\arguments{
\item{year}{The year of the census.}
}
\value{
A character vector of file names found.
}
\description{
Lists the data files available in a downloaded Higher Education Census.
Useful for exploring the contents of the ZIP file.
}
\examples{
\dontrun{
list_censo_superior_files(2023)
}
}
\seealso{
Other Higher Education Census functions: 
\code{\link{get_censo_superior}()}
}
\concept{Higher Education Census functions}
