% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ideb.R
\name{list_ideb_available}
\alias{list_ideb_available}
\title{List available IDEB data}
\usage{
list_ideb_available()
}
\value{
A tibble with available IDEB datasets.
}
\description{
Lists the IDEB data files available in the INEP portal.
}
\examples{
list_ideb_available()
}
\seealso{
Other IDEB functions: 
\code{\link{get_ideb}()},
\code{\link{get_ideb_series}()}
}
\concept{IDEB functions}
