% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ideb.R
\name{read_ideb_excel}
\alias{read_ideb_excel}
\title{Read IDEB Excel file}
\usage{
read_ideb_excel(file)
}
\arguments{
\item{file}{Path to the Excel file.}
}
\value{
A tibble with the data.
}
\description{
Internal function to read IDEB Excel files.
}
\keyword{internal}
